fix(metal): Enable ARC and fix MSL double precision issues for macOS

+++ backend/src/kernels/metal/gate_kernels.metal
@@ -2,8 +2,8 @@
 using namespace metal;
 
 struct Complex {
-    double real;
-    double imag;
+    float real;
+    float imag;
 };
 
 // Simple Hadamard Gate Kernel for Metal
